+++ lib/src/widget/countries_scope.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_simple_country_picker/flutter_simple_country_picker.dart';
 import 'package:flutter_simple_country_picker/src/controller/countries_controller.dart';
+import 'package:flutter_simple_country_picker/src/controller/countries_provider.dart';
 import 'package:flutter_simple_country_picker/src/controller/countries_state.dart';
 import 'package:meta/meta.dart';