+++ test/widget_test/flutter_simple_country_picker_test.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  testWidgets('Example_first_widget_test -', (tester) async {});
-}