+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.0.1-pre.3
+- **ADDED**: Selected country to show mark intro list of countries
 - **FIXED**: `Android`
 - **FIXED**: `ios`